# Add benchmark script